Refactor to enable upload to happen in s3backup package
