+++ cmd/root.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"path"
@@ -65,7 +66,7 @@ func doUpload(cmd *cobra.Command, args []string) {
 	store := createStore(config.S3)
 	remoteIndex := readRemoteIndex(config, store)
 	localIndex := createLocalIndex()
-	uploadDifferences(localIndex, remoteIndex, store)
+	s3backup.UploadDifferences(localIndex, remoteIndex, store, getFile)
 	uploadIndex(localIndex, store)
 
 	doLog("Finished")
@@ -135,26 +136,15 @@ func createLocalIndex() *s3backup.Index {
 	return localIndex
 }
 
-func uploadDifferences(localIndex, remoteIndex *s3backup.Index, store *s3.Store) {
-	diff := localIndex.Diff(remoteIndex)
-	for p, srcFile := range diff.Files {
-		r, err := os.Open(path.Clean(p))
-		defer func() {
-			_ = r.Close()
-		}()
+func getFile(p string) io.ReadCloser {
+	r, err := os.Open(path.Clean(p))
 
-		if err != nil {
-			fmt.Fprintln(os.Stderr, err.Error())
-			os.Exit(1)
-		}
-
-		doLog("Uploading %s as %s\n", p, srcFile.Key)
-		err = store.Save(srcFile.Key, r)
-		if err != nil {
-			fmt.Fprintln(os.Stderr, err.Error())
-			os.Exit(1)
-		}
+	if err != nil {
+		fmt.Fprintln(os.Stderr, err.Error())
+		os.Exit(1)
 	}
+
+	return r
 }
 
 func uploadIndex(index *s3backup.Index, store *s3.Store) {
